Enhance WebSocketController to handle marketplace product deletion and improve null checks for direct chats and groups

+++ src/main/java/com/vibez/engine/WebSocket/WebSocketController.java
@@ -24,6 +24,7 @@ import com.vibez.engine.Model.User;
 import com.vibez.engine.Model.UserUpdate;
 import com.vibez.engine.Repository.DirectChatRepo;
 import com.vibez.engine.Repository.GroupRepo;
+import com.vibez.engine.Repository.MarketplaceRepo;
 import com.vibez.engine.Repository.UserRepo;
 import com.vibez.engine.Service.DirectChatService;
 import com.vibez.engine.Service.FriendshipService;
@@ -64,6 +65,9 @@ public class WebSocketController implements WebSocketHandler {
 
     @Autowired
     private DirectChatRepo directChatRepo;
+
+    @Autowired
+    private MarketplaceRepo MarketplaceRepo;
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -302,31 +306,41 @@ public class WebSocketController implements WebSocketHandler {
         String uniqeId = "message_" + System.currentTimeMillis();
         String typeOfAction = null;
         User existingUser = userRepo.findByUserId(user.getUserId());
-
+        List <Marketplace> products = MarketplaceRepo.findBySellerId(user.getUserId());
         List <String> directChats = existingUser.getDirectChatIds();
         List <String> groups = existingUser.getGroupIds();
-        for (String directChatId : directChats) {
-            DirectChat  directChat = directChatRepo.findByChatId(directChatId);
-            List <String> memberIds = directChat.getMemberIds();
-            String otherUser = memberIds.stream()
-                                    .filter(id -> !id.equals(user.getUserId()))
-                                    .findFirst()
-                                    .orElse(null); 
-            User otherUserObj = userRepo.findByUserId(otherUser);
-            otherUserObj.getDirectChatIds().remove(directChatId);
-            userRepo.save(otherUserObj);
-            directChatRepo.deleteById(directChatId);
+        
+        if(directChats != null){
+            for (String directChatId : directChats) {
+                DirectChat  directChat = directChatRepo.findByChatId(directChatId);
+                List <String> memberIds = directChat.getMemberIds();
+                String otherUser = memberIds.stream()
+                                        .filter(id -> !id.equals(user.getUserId()))
+                                        .findFirst()
+                                        .orElse(null); 
+                User otherUserObj = userRepo.findByUserId(otherUser);
+                otherUserObj.getDirectChatIds().remove(directChatId);
+                userRepo.save(otherUserObj);
+                directChatRepo.deleteById(directChatId);
 
-            typeOfAction = "directChat";
+                typeOfAction = "directChat";
+            }
         }
-        for (String groupId : groups) {
-            Groups group = groupRepo.findByGroupId(groupId);
-            group.getMemberIds().remove(user.getUserId());
-            groupRepo.save(group);
-            typeOfAction = "groupChat";
+        if (groups != null){
+            for (String groupId : groups) {
+                Groups group = groupRepo.findByGroupId(groupId);
+                group.getMemberIds().remove(user.getUserId());
+                groupRepo.save(group);
+                typeOfAction = "groupChat";
+            }
         }
-        
-        
+        if (products != null){
+            for (Marketplace product : products) {
+                MarketplaceRepo.deleteById(product.getProductId());
+                typeOfAction = "marketplace";
+            }
+        }
+
         userRepo.delete(existingUser);
         Map<String, Object> message = new HashMap<>();
         message.put("id", uniqeId);
